+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **Version v2 (master)**
 
+- Scale STATPOP to projected household-level control totals
 - Fix municipality type and canton imputation
 - Write out detailed information from MZ for PT trips
 - Fix education location